+++ pac4j-core/src/main/java/org/pac4j/core/util/CommonHelper.java
@@ -343,4 +343,8 @@ public final class CommonHelper {
         }
         return constructor;
     }
+
+    public static String ifBlank(final String value, final String defaultValue) {
+        return isBlank(value) ? defaultValue : value;
+    }
 }