+++ pac4j-saml-opensamlv5/src/main/java/org/pac4j/saml/metadata/SAML2MetadataGenerator.java
@@ -2,18 +2,17 @@ package org.pac4j.saml.metadata;
 
 import org.opensaml.saml.metadata.resolver.MetadataResolver;
 import org.opensaml.saml.saml2.metadata.EntityDescriptor;
-import org.springframework.core.io.Resource;
 
 /**
  * Builds metadata and the relevant resolvers.
  * @author Misagh Moayyed
  */
 public interface SAML2MetadataGenerator {
-    MetadataResolver buildMetadataResolver(Resource metadataResource) throws Exception;
+    MetadataResolver buildMetadataResolver() throws Exception;
 
     String getMetadata(EntityDescriptor entityDescriptor) throws Exception;
 
     EntityDescriptor buildEntityDescriptor();
 
-    boolean storeMetadata(String metadata, Resource resource, boolean force) throws Exception;
+    boolean storeMetadata(String metadata, boolean force) throws Exception;
 }